add db logic

+++ java/todo/src/main/java/com/interco/domain/ports/TaskRepository.java
@@ -19,5 +19,5 @@ public interface TaskRepository {
 
     void deleteAll();
 
-    Integer count();
+    long count();
 }
